Introduce a workloads watcher
Adds a new watcher that uses `cache.Informer` instances, one for each
watched namespace, to watch for `Service` resources. The watcher then
filters out services that cannot be intercepted.

The watcher is primarily intended for a new WatchWorkloads gRPC method
but the cache of currently interceptable services can also be used for
speeding up the `list` command.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>